requirements.txt
django=='*'
djangorestframework = '*'